Pull up and reuse bulk indexing method
